Fix a couple trucation warnings
These were showing up in the on-Linux-in-Fuchsia-tree build. It seemed
easier to just cast these two locations than attempt to suppress the
warnings in this case (vs.
https://chromium-review.googlesource.com/c/crashpad/crashpad/+/2411131
where they're being disabled in Crashpad-proper).

Bug: fuchsia:58162

+++ base/logging.cc
@@ -362,7 +362,7 @@ void LogMessage::Init(const char* function) {
 #elif defined(OS_ANDROID)
   pid_t thread = gettid();
 #elif defined(OS_LINUX)
-  pid_t thread = syscall(__NR_gettid);
+  pid_t thread = static_cast<int>(syscall(__NR_gettid));
 #elif defined(OS_WIN)
   DWORD thread = GetCurrentThreadId();
 #elif defined(OS_FUCHSIA)
